+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,32 @@
 # react-example
 
+## 2.8.56
+
+### Patch Changes
+
+- @stackframe/react@2.8.56
+
+## 2.8.55
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/react@2.8.55
+
+## 2.8.54
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/react@2.8.54
+
+## 2.8.53
+
+### Patch Changes
+
+- @stackframe/react@2.8.53
+
 ## 2.8.52
 
 ### Patch Changes