+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/stack-sc
 
+## 2.8.56
+
+## 2.8.55
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.54
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.53
+
 ## 2.8.52
 
 ## 2.8.51