+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @stackframe/stack-shared
 
+## 2.8.56
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.55
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.54
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.53
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.52
 
 ### Patch Changes