+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @stackframe/stack-ui
 
+## 2.8.56
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.56
+
+## 2.8.55
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.55
+
+## 2.8.54
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.54
+
+## 2.8.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.53
+
 ## 2.8.52
 
 ### Patch Changes